added foundry console logs and renamed the rule
