+++ lib/rules/best-practises/index.js
@@ -6,7 +6,7 @@ const NoEmptyBlocksChecker = require('./no-empty-blocks')
 const NoUnusedVarsChecker = require('./no-unused-vars')
 const PayableFallbackChecker = require('./payable-fallback')
 const ReasonStringChecker = require('./reason-string')
-const NoConsoleLogChecker = require('./no-console-log')
+const NoConsoleLogChecker = require('./no-console')
 
 module.exports = function checkers(reporter, config, inputSrc) {
   return [
