+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "solhint",
-  "version": "3.3.7",
+  "version": "3.3.8",
   "description": "Solidity Code Linter",
   "main": "lib/index.js",
   "keywords": [
